pulled out extra column margins
