+++ app/page.tsx
@@ -3,12 +3,12 @@ import Link from 'next/link'
 import { Suspense } from 'react'
 // import Table from '@/components/table'
 import TablePlaceholder from '@/components/table-placeholder'
-import TableDirectory from '@/components/table-directory'
-import Filters from '@/components/filters'
+import Directory from '@/components/listing/directory'
+import Filters from '@/components/listing/filters'
 import Episodes from '@/components/listing/episodes'
-import Logo from '@/components/header/logo'
 import MenuLeft from '@/components/header/menu-left'
 import MenuRight from '@/components/header/menu-right'
+import About from '@/components/footer/about'
 
 // Prisma does not support Edge without the Data Proxy currently
 // export const runtime = 'edge'
@@ -18,7 +18,7 @@ export const dynamic = 'force-dynamic'
 export default function Home() {
   return (
     <main className="pattern">
-      <div className="container">
+      <div className="wrapper">
       
       <div className="logo"></div>
       <MenuLeft />
@@ -36,47 +36,11 @@ export default function Home() {
             {/* @ts-expect-error Async Server Component */}
             <Filters />
             {/* @ts-expect-error Async Server Component */}
-            <TableDirectory />          
+            <Directory />          
         </Suspense>
-        
       
+      <About />
 
-
-{/* Footer content */}
-        {/* <div className="flex justify-center space-x-5 pt-10 mt-10 border-t border-gray-300 w-full max-w-xl text-gray-600"> */}
-          
-          {/* <Link
-            href="https://postgres-drizzle.vercel.app/"
-            className="font-medium underline underline-offset-4 hover:text-black transition-colors"
-          >
-            Drizzle
-          </Link>
-        </div>
-
-        <div className="sm:absolute sm:bottom-0 w-full px-20 py-10 flex justify-between">
-          <Link href="https://vercel.com">
-            <Image
-              src="/vercel.svg"
-              alt="Vercel Logo"
-              width={100}
-              height={24}
-              priority
-            />
-          </Link>
-          <Link
-            href="https://github.com/vercel/examples/tree/main/storage/postgres-prisma"
-            className="flex items-center space-x-2"
-          >
-            <Image
-              src="/github.svg"
-              alt="GitHub Logo"
-              width={24}
-              height={24}
-              priority
-            />
-            <p className="font-light">Source</p>
-          </Link> */}
-        {/* </div> */}
       </div>
     </main>
   )
