+++ components/footer/about.tsx
@@ -1,0 +1,63 @@
+
+export default function About() {
+    return (
+
+<div className="aboutContent">
+    <p>
+    This is a phonebook for you and your creative peers in the Pacific Northwest. We also have an interview series called the Seattle Creative Show that can either be found on the homepage, in your favorite podcast player or <a href="https://seattlecreative.show/">see all available platforms here.</a> If you have any questions or suggestions please feel free to reach out <a href="mailto:jonny@seattlecreative.directory" target="_blank">in email.</a>
+    </p>
+    <p>
+    🔎 For more behind the scenes, there's a real nice <a href="http://marketingnw.com/new-seattle-creative-directory/">write up in Marketing Northwest on the project</a> that you can check out.
+    </p>
+    <p>
+    📓 To submit a new entry, please answer <a href="https://forms.gle/5rwbdmCDipjwtBJY8" target="_blank">a few questions</a> about your creative business.
+    </p>
+    <p>
+    Created with love by<br />
+    <a href="https://twitter.com/jonmccon" target="_blank">Jonny McConnell</a><br />
+    <a href="https://www.instagram.com/ben_busy" target="_blank">Ben Hubbard</a><br />
+    <a href="https://sea-to-sound.com" target="_blank">Lee Gehrig</a><br />
+    <a href="https://github.com/jonmccon/seattle-creative-directory">Source</a>
+    </p>
+
+</div>
+
+
+
+// {/* Footer content */}
+//         {/* <div className="flex justify-center space-x-5 pt-10 mt-10 border-t border-gray-300 w-full max-w-xl text-gray-600"> */}
+          
+//           {/* <Link
+//             href="https://postgres-drizzle.vercel.app/"
+//             className="font-medium underline underline-offset-4 hover:text-black transition-colors"
+//           >
+//             Drizzle
+//           </Link>
+//         </div>
+
+//         <div className="sm:absolute sm:bottom-0 w-full px-20 py-10 flex justify-between">
+//           <Link href="https://vercel.com">
+//             <Image
+//               src="/vercel.svg"
+//               alt="Vercel Logo"
+//               width={100}
+//               height={24}
+//               priority
+//             />
+//           </Link>
+//           <Link
+//             href="https://github.com/vercel/examples/tree/main/storage/postgres-prisma"
+//             className="flex items-center space-x-2"
+//           >
+//             <Image
+//               src="/github.svg"
+//               alt="GitHub Logo"
+//               width={24}
+//               height={24}
+//               priority
+//             />
+//             <p className="font-light">Source</p>
+//           </Link> */}
+//         {/* </div> */}
+
+        )}
